Access jekyll config via singleton to shut up all the writing to STDOUT
So, as we know, accessing `Jekyll.configuration` will result in some logging to
STDOUT. Honestly I don't know why, but it is what it is.

One way to reduce the number of times I'd previously accessed the config, is to assign
and memoize it, once, to a singleton. The ImgproxyConfig instance can now access this
via it's eigenclass. Easy enough. Delegating `#imgproxy_config` to `self.class` now
will work as expected.

Also the spec now just overwrites the class variable with a nil and lets it re-initialize
itself inside the helper method that stubs it.