Refactor how Jekyll config is accessed
Prior to this commit, Jekyll's configuration was being accessed directly via the
Jekyll.configuration singleton. Turns out this is unnecessary, considering we already have an object
at our disposal, the `context` object that's passed to the `render` instance method of the
`Liquid::Tag` subclass.

So, let's use that instead of accessing the config all over again. No need.